+++ config/config.exs
@@ -56,7 +56,7 @@ config :esbuild,
 
 # https://fly.io/phoenix-files/tailwind-standalone/
 config :tailwind,
-  version: "3.2.1",
+  version: "3.2.4",
   default: [
     args: ~w(
       --config=tailwind.config.js