+++ lib/lightning_web.ex
@@ -43,10 +43,13 @@ defmodule LightningWeb do
     end
   end
 
-  def live_view do
+  def live_view(opts \\ []) do
     quote do
-      use Phoenix.LiveView,
-        layout: {LightningWeb.LayoutView, :live}
+      @opts Keyword.merge(
+              [layout: {LightningWeb.LayoutView, :live}],
+              unquote(opts)
+            )
+      use Phoenix.LiveView, @opts
 
       unquote(view_helpers())
     end
@@ -121,6 +124,10 @@ defmodule LightningWeb do
   @doc """
   When used, dispatch to the appropriate controller/view/etc.
   """
+  defmacro __using__({which, opts}) when is_atom(which) do
+    apply(__MODULE__, which, [opts])
+  end
+
   defmacro __using__(which) when is_atom(which) do
     apply(__MODULE__, which, [])
   end