+++ lib/lightning_web/live/job_live/manual_run_component.ex
@@ -7,6 +7,7 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
   attr :job_id, :string, required: true
   attr :current_user, Lightning.Accounts.User, required: true
   attr :builder_state, :any, required: true
+  attr :on_run, :any, required: true
   attr :selected_dataclip, Lightning.Invocation.Dataclip
 
   @impl true
@@ -30,14 +31,13 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
           phx-change="changed"
           phx-target={@myself}
         />
-        <pre class="flex-1 bg-gray-100 m-2 p-3 font-mono"><%= @selected_dataclip.body %></pre>
+        <LightningWeb.RunLive.Components.log_view log={@selected_dataclip.body} />
         <Common.button
           text="Run"
           disabled={!@changeset.valid?}
           phx-click="confirm"
           phx-target={@myself}
         />
-        <.live_info_block myself={@myself} flash={@flash} />
       </.form>
     </div>
     """
@@ -55,10 +55,11 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
   @impl true
   def update(
         %{
-          job_id: job_id,
+          builder_state: builder_state,
           current_user: current_user,
           id: id,
-          builder_state: builder_state
+          job_id: job_id,
+          on_run: on_run
         },
         socket
       ) do
@@ -97,22 +98,24 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
        builder_state: builder_state,
        dataclips: dataclips,
        dataclips_options: dataclips_options,
+       on_run: on_run,
        selected_dataclip: selected_dataclip |> format()
      )
      |> update_form(init_form)}
   end
 
   @impl true
-
   def handle_event("confirm", _params, socket) do
     socket.assigns.changeset
     |> Ecto.Changeset.put_change(:user, socket.assigns.current_user)
     |> create_manual_workorder()
     |> case do
-      {:ok, _} ->
+      {:ok, %{attempt_run: attempt_run}} ->
+        socket.assigns.on_run.(attempt_run)
+
         {:noreply,
          socket
-         |> put_flash(:info, "Run enqueued.")}
+         |> push_event("push-hash", %{hash: "output"})}
 
       {:error, changeset} ->
         {:noreply,
@@ -142,7 +145,7 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
   end
 
   defp format(dataclip) when is_nil(dataclip) do
-    %{id: "", body: ""}
+    %{id: "", body: []}
   end
 
   defp format(dataclip) do
@@ -152,6 +155,7 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
         dataclip.body
         |> Jason.encode!()
         |> Jason.Formatter.pretty_print()
+        |> String.split("\n")
     }
   end
 
@@ -181,12 +185,16 @@ defmodule LightningWeb.JobLive.ManualRunComponent do
     with {:ok, dataclip} <- get_dataclip(changeset),
          {:ok, job} <- get_job(changeset),
          user <- changeset |> Ecto.Changeset.get_field(:user) do
-      {:ok, %{attempt_run: attempt_run}} =
-        Lightning.WorkOrderService.multi_for_manual(job, dataclip, user)
-        |> Repo.transaction()
+      # HACK: Oban's testing functions only apply to `self` and LiveView
+      # tests run in child processes, so for now we need to set the testing
+      # mode from within the process.
+      Process.put(:oban_testing, :manual)
 
-      Lightning.Pipeline.new(%{attempt_run_id: attempt_run.id})
-      |> Oban.insert()
+      Lightning.WorkOrderService.multi_for_manual(job, dataclip, user)
+      |> Oban.insert(:pipeline, fn %{attempt_run: attempt_run} ->
+        Lightning.Pipeline.new(%{attempt_run_id: attempt_run.id})
+      end)
+      |> Repo.transaction()
     end
   end